fetch/metadata/generated/serviceworker.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/serviceworker.https.sub.html
-->
<!DOCTYPE html>
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for Service Workers</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(t, url, options, event, clear) {
    // Register a service worker and check the request header.
    return navigator.serviceWorker.register(url, options)
      .then((registration) => {
        t.add_cleanup(() => registration.unregister());
        if (event === 'register') {
          return;
        }
        return clear().then(() => registration.update());
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(t, url, {}, 'register')
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
        });
  }, 'sec-fetch-frame-ancestors - Same origin, no options - registration');

  promise_test((t) => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(t, url, {}, 'update', () => retrieve(key))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
        });
  }, 'sec-fetch-frame-ancestors - Same origin, no options - updating');
  </script>
</body>
</html>
